+++ modernRX/dataset.hpp
@@ -2,7 +2,7 @@
 
 
 /*
-* Multi-threaded and AVX2 enhanced implementation of RandomX Dataset: https://github.com/tevador/RandomX/blob/master/doc/specs.md#7-dataset
+* Multi-threaded, JIT-compiled and AVX2 enhanced implementation of RandomX Dataset: https://github.com/tevador/RandomX/blob/master/doc/specs.md#7-dataset
 * This is used as read-only memory by RandomX programs to calculate hashes.
 */
 
@@ -14,7 +14,7 @@
 
 namespace modernRX {
 
-    // Fills read-only memory used by RandomX programs to calculate hashes according to https://github.com/tevador/RandomX/blob/master/doc/specs.md#7-dataset.
-    // Needs cache as an argon2d filled memory buffer and 8 superscalar programs.
+    // Compiles superscalar programs and fills read-only memory used by RandomX programs to calculate hashes according to https://github.com/tevador/RandomX/blob/master/doc/specs.md#7-dataset.
+    // Needs cache as an Argon2d filled memory buffer and 8 superscalar programs.
     [[nodiscard]] HeapArray<DatasetItem, 4096> generateDataset(const_span<argon2d::Block> cache, const_span<SuperscalarProgram, Rx_Cache_Accesses> programs);
 }