+++ modernRX/virtualmem.hpp
@@ -16,28 +16,35 @@
 #include "aliases.hpp"
 
 namespace modernRX {
+    template<typename T>
+    inline constexpr bool is_vector_v = std::same_as<T, std::vector<typename T::value_type, typename T::allocator_type> >;
+
+    template<typename T>
+    inline constexpr bool is_spanable_v = std::is_constructible_v<const_span<typename T::value_type>, T>;
+
     // Allocates executable memory and returns a function pointer to it.
     // Type of the function pointer is specified by the template parameter.
     // The function pointer is wrapped in a unique_ptr with a custom deleter that will free allocated memory and data associated with program at destruction.
     // May throw if memory allocation or protection fails.
-    // Deleter dont throw on failure.
     // 
-    // Function takes data as a parameter and prolongs its lifetime until the JIT-compiled program is destroyed.
+    // Function takes data as a parameter and prolongs its lifetime until the JIT-compiled function is destroyed.
     // Somewhat hacky, would be nice to find a better solution.
-    template<typename Fn, typename Byte = std::byte>
-    requires (sizeof(Byte) == 1 && std::is_trivially_copyable_v<Byte>)
-    jit_function_ptr<Fn> makeExecutable(const_span<Byte> code, std::vector<Byte>&& data) {
+    template<typename Fn, typename Code, typename Data>
+    requires is_spanable_v<Code> && is_vector_v<Data>
+    jit_function_ptr<Fn> makeExecutable(Code code, Data&& data) {
+        const auto code_size{ const_span<typename Code::value_type>(code).size_bytes() };
+
         // Alloc buffer for writing code.
-        auto buffer{ VirtualAlloc(nullptr, code.size_bytes(), MEM_COMMIT | MEM_RESERVE, PAGE_READWRITE) };
+        auto buffer{ VirtualAlloc(nullptr, code_size, MEM_COMMIT | MEM_RESERVE, PAGE_READWRITE) };
         if (buffer == nullptr) {
             throw std::runtime_error(std::format("Failed to allocate memory with error: {:d}", GetLastError()));
         }
 
-        std::memcpy(buffer, code.data(), code.size_bytes());
+        std::memcpy(buffer, code.data(), code_size);
 
         // Protect from writing, but make code executable.
         DWORD dummy{};
-        if (!VirtualProtect(buffer, code.size_bytes(), PAGE_EXECUTE_READ, &dummy)) {
+        if (!VirtualProtect(buffer, code_size, PAGE_EXECUTE_READ, &dummy)) {
             VirtualFree(buffer, 0, MEM_RELEASE);
             throw std::runtime_error(std::format("Failed to protect memory with error: {:d}", GetLastError()));
         }